Add utility functions for file system operations
Introduced new module fs.js with multiple utility functions to handle file system operations such as creating directories, renaming folders, downloading files, and extracting zip files. These operations are essential for various process like setting up environments, downloading dependencies or cleaning up temporary files.
