+++ lib/utils.js
@@ -1,8 +1,24 @@
 const path = require('path');
 const fs = require('fs');
 const { exec } = require('node:child_process');
-const extract = require('extract-zip');
-const https = require('node:https');
+
+/**
+ * Reads a PHP file, extracts, and returns the WordPress version number.
+ *
+ * @param {string} fileContent - Path to the PHP file to read
+ * @param variableName - The name of the variable to search adn replace
+ * @return {string|null} WordPress version number or null if not found or in case of an error
+ */
+function geVarFromPHPFile (fileContent, variableName = '$wp_version') {
+  // Define a regular expression to match the variableName and its value with both the single and double quotes
+  const versionRegex = new RegExp(`${variableName}\\s*=\\s*['"]([^'"]+)['"]`, 'g');
+
+  // Use the regular expression to extract the version number
+  const match = fileContent.match(versionRegex);
+
+  // Return the version number or null if not found
+  return match ? match[1] : null;
+}
 
 /**
  * Reads wp-package.json from the root folder and extracts the value of the --template option or uses the default.
@@ -30,41 +46,6 @@ function getConfig (args) {
   }
 
   return config;
-}
-
-/**
- * Create a temporary directory if it does not already exist.
- */
-function makeDir (dirpath) {
-  if (!fs.existsSync(dirpath)) {
-    fs.mkdirSync(dirpath, { recursive: true });
-  }
-}
-
-/**
- * Asynchronously cleans up a temporary directory.
- *
- * @param {string} dir - The path to the temporary directory.
- * @return {void} A promise that resolves when the cleanup is complete.
- */
-async function cleanup (dir) {
-  try {
-    fs.rmSync(dir, { recursive: true });
-    console.log(`🧹 ${dir} removed successfully.`);
-  } catch (err) {
-    // File deletion failed
-    console.error(err.message);
-  }
-}
-
-/**
- * Renames a folder from the old path to the new path.
- *
- * @param {string} oldPath - The path of the folder to be renamed.
- * @param {string} newPath - The new path of the folder.
- */
-function renameFolder (oldPath, newPath) {
-  fs.renameSync(oldPath, newPath);
 }
 
 /**
@@ -102,88 +83,6 @@ function getDownloadUrl (packageName, packageVersion, type) {
 
   // otherwise we assume it's a repo on WordPress.org
   return `https://downloads.wordpress.org/${type}/${packageName}.zip`;
-}
-
-/**
- * Downloads a file from the specified URL and saves it to the target file.
- *
- * @param {string} url - The URL of the file to download.
- * @param {string} targetFile - The file path where the downloaded file will be saved.
- * @return {Promise<void>} A promise that resolves when the file is successfully downloaded and saved, or rejects with an error if there was an issue.
- */
-async function downloadFile (url, targetFile) {
-  if (fs.existsSync(targetFile)) {
-    console.log(`ℹ️ ${targetFile} already exists. Skipping download.`);
-    return;
-  }
-  try {
-    return await new Promise((resolve, reject) => {
-      https.get(
-        url,
-        { headers: { 'User-Agent': 'nodejs' } },
-        async (response) => {
-          const code = response.statusCode ?? 0;
-
-          if (code >= 400) {
-            return reject(new Error(response.statusMessage));
-          }
-
-          if (code > 300 && code < 400 && !!response.headers.location) {
-            return resolve(await downloadFile(response.headers.location, targetFile));
-          }
-
-          const fileWriter = fs.createWriteStream(targetFile).on('finish', () => {
-            resolve({});
-          });
-
-          response.pipe(fileWriter);
-        }).on('error', (error) => {
-        reject(error);
-      });
-    });
-  } catch (error) {
-    throw new Error(error);
-  }
-}
-
-/**
- * Extracts a zip file to a target directory.
- *
- * @param {string} zipFilePath - The path of the zip file to extract.
- * @param {string} targetDirectory - The directory to extract the zip file to.
- * @return {Promise<string>} Returns true if the extraction is successful, false otherwise.
- */
-async function extractZip (zipFilePath, targetDirectory) {
-  let commonRootPath; // Variable to store the common root path
-
-  try {
-    await extract(zipFilePath, {
-      dir: targetDirectory,
-      onEntry: (entry) => {
-        const entryPathParts = entry.fileName.split('/');
-
-        if (!commonRootPath) {
-          // Initialize the common root path with the first entry
-          commonRootPath = entryPathParts[0];
-        } else {
-          // Update the common root path based on the current entry
-          for (let i = 0; i < entryPathParts.length; i++) {
-            if (commonRootPath.split('/')[i] !== entryPathParts[i]) {
-              commonRootPath = commonRootPath.split('/').slice(0, i).join('/');
-              break;
-            }
-          }
-        }
-      }
-    });
-
-    // Return the root folder name
-    console.log(`📂 Extracted to ${commonRootPath}`);
-    return commonRootPath;
-  } catch (err) {
-    console.error(`📛 Error extracting ${zipFilePath} zip: ${err}`);
-    return err;
-  }
 }
 
 /**
@@ -300,20 +199,26 @@ async function isWPCLIAvailable () {
   }
 }
 
+function printTimePassed (startTime) {
+  // End the timer
+  const endTime = Date.now();
+
+  // Calculate the time passed
+  const timePassed = endTime - startTime;
+  console.log(`🕒 Time passed: ${timePassed} milliseconds`);
+}
+
 module.exports = {
+  geVarFromPHPFile,
   getConfig,
-  makeDir,
-  cleanup,
-  renameFolder,
-  downloadFile,
   getWordPressDownloadUrl,
   getDownloadUrl,
-  extractZip,
   installNpmPackages,
   installComposer,
   replaceDbConstant,
   replaceDbConstantBool,
   generateSalt,
   replaceEmptySalts,
-  isWPCLIAvailable
+  isWPCLIAvailable,
+  printTimePassed
 };
